Integration-2 state-wise data + FE modifications
