+++ todo_drf/frontend/templates/frontend/prediction_V2_V0.html
@@ -110,7 +110,7 @@
         <li><a href="{% url "frontend:start_disp" %}">Stream & Check</a></li>
         <li><a href="{% url "frontend:frame_4" %}">Monitor Web</a></li>
         <li><a href="{% url "frontend:dashboard_index" %}">Dashboard</a></li>
-        <li><a href="{% url "frontend:trace_1" %}">Tracing Fake News</a></li>
+        <li><a href="{% url "frontend:trace_landing" %}">Tracing Fake News</a></li>
         <li><a href="{% url "frontend:render_manual_page" %}">Manual Check</a></li>
 
       </ul>
